Add snapshot fix
