+++ src/routers/daoRouter.jsx
@@ -37,6 +37,7 @@ import ProposalTypes from '../pages/ProposalTypes';
 import ProposalWatcher from '../pages/ProposalWatcher';
 import Settings from '../pages/Settings';
 import Snapshot from '../pages/Snapshot';
+import SnapshotSettings from '../pages/SnapshotSettings';
 import SuperfluidMinion from '../pages/SuperfluidMinion';
 import ThemeBuilder from '../pages/ThemeBuilder';
 import Treasury from '../pages/Treasury';
@@ -269,6 +270,12 @@ const DaoRouter = () => {
         </Route>
         <Route exact path={`${path}/boost/mintgate`}>
           <MintGate daoMetaData={daoMetaData} />
+        </Route>
+        <Route exact path={`${path}/boost/snapshot/settings`}>
+          <SnapshotSettings
+            daoMetaData={daoMetaData}
+            refetchMetaData={refetchMetaData}
+          />
         </Route>
         <Route exact path={`${path}/boost/snapshot`}>
           <Snapshot
